ugc/src/requirements.txt
clickhouse-driver==0.2.6
kafka-python==2.0.2
pydantic[dotenv]==1.10.5
fastapi==0.94.1
gunicorn==20.1.0
orjson==3.5.1
uvicorn==0.12.2
redis==4.4.2
PyJWT==2.7.0
uvloop==0.17.0; sys_platform != "win32" and implementation_name == "cpython"
httptools==0.5.0
backoff==2.2.1